Reviewed-on: https://neptune.builders/core-team/twenty-first/pulls/89

Redefines a B field element to be a 64-bit number so some calculations (add, sub, neg) can be done with only 64-bit registers.

This speeds up a single RP digest calculation from 80µs to 20µs. And 2^14 RP pair calculations from 1700ms to 429ms. This is a 50 % speedup compared to what winterfell are doing. The weird optimization on `add`, involving an extra check that `val > BFieldElement::MAX` makes this implementation 50 % faster than novifinancial/winterfell.

+++ src/shared_math/polynomial.rs
@@ -456,7 +456,7 @@ impl<PFElem: PrimeField> Polynomial<PFElem> {
 
         let result_degree: u64 = 2 * self.degree() as u64;
         let order = roundup_npo2(result_degree + 1);
-        let (root_res, _) = self.coefficients[0].get_primitive_root_of_unity(order as u128);
+        let (root_res, _) = self.coefficients[0].get_primitive_root_of_unity(order);
         let root = match root_res {
             Some(n) => n,
             None => panic!("Failed to find primitive root for order = {}", order),
